+++ backend/tests/Designer.Tests/Controllers/PreviewController/GetInstanceIdTests.cs
@@ -1,7 +1,22 @@
-﻿namespace Designer.Tests.Controllers.PreviewController
+﻿using System;
+using System.Net;
+using System.Net.Http;
+using System.Text.Json;
+using System.Threading.Tasks;
+using Altinn.Platform.Storage.Interface.Models;
+using Microsoft.AspNetCore.Mvc.Testing;
+using Newtonsoft.Json;
+using Xunit;
+
+namespace Designer.Tests.Controllers.PreviewController
 {
-    public class GetInstanceIdTests
+    public class GetInstanceIdTests : PreviewControllerTestsBase<GetInstanceIdTests>
     {
-        
+
+        public GetInstanceIdTests(WebApplicationFactory<Altinn.Studio.Designer.Controllers.PreviewController> factory) : base(factory)
+        {
+        }
+
+
     }
 }
