+++ backend/tests/Designer.Tests/Controllers/PreviewController/InstanceForNextTaskTests.cs
@@ -4,28 +4,24 @@ using System.Net.Http;
 using System.Text.Json;
 using System.Threading.Tasks;
 using Altinn.Platform.Storage.Interface.Models;
-using Designer.Tests.Utils;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Newtonsoft.Json;
 using Xunit;
 
 namespace Designer.Tests.Controllers.PreviewController
 {
-    public class InstancesTests : PreviewControllerTestsBase<InstancesTests>
+    public class InstanceForNextTaskTests : PreviewControllerTestsBase<InstanceForNextTaskTests>
     {
 
-        public InstancesTests(WebApplicationFactory<Altinn.Studio.Designer.Controllers.PreviewController> factory) : base(factory)
+        public InstanceForNextTaskTests(WebApplicationFactory<Altinn.Studio.Designer.Controllers.PreviewController> factory) : base(factory)
         {
         }
 
         [Fact]
-        public async Task Post_Instance_Ok()
+        public async Task Get_InstanceForNextProcess_Ok()
         {
-            string targetRepository = TestDataHelper.GenerateTestRepoName();
-            await CopyRepositoryForTest(Org, App, Developer, targetRepository);
-
-            string dataPathWithData = $"{Org}/{targetRepository}/instances?instanceOwnerPartyId=51001";
-            using HttpRequestMessage httpRequestMessage = new(HttpMethod.Post, dataPathWithData);
+            string dataPathWithData = $"{Org}/{App}/instances/{PartyId}/{InstanceGuId}";
+            using HttpRequestMessage httpRequestMessage = new(HttpMethod.Get, dataPathWithData);
             httpRequestMessage.Headers.Referrer = new Uri($"{MockedReferrerUrl}?org={Org}&app={App}&selectedLayoutSetInEditor=");
 
             using HttpResponseMessage response = await HttpClient.Value.SendAsync(httpRequestMessage);
@@ -34,18 +30,15 @@ namespace Designer.Tests.Controllers.PreviewController
             string responseBody = await response.Content.ReadAsStringAsync();
             JsonDocument responseDocument = JsonDocument.Parse(responseBody);
             Instance instance = JsonConvert.DeserializeObject<Instance>(responseDocument.RootElement.ToString());
-            Assert.Equal("test-datatask-id", instance.Data[0].Id);
-            Assert.Equal("Task_1", instance.Process.CurrentTask.ElementId);
+            Assert.Equal($"{PartyId}/{InstanceGuId}", instance.Id);
+            Assert.Equal("ttd", instance.Org);
         }
 
         [Fact]
-        public async Task Post_InstanceForStatefulApp_Ok()
+        public async Task Get_InstanceForNextTaskForStatefulApp_Ok_TaskIsIncreased()
         {
-            string targetRepository = TestDataHelper.GenerateTestRepoName();
-            await CopyRepositoryForTest(Org, StatefulApp, Developer, targetRepository);
-
-            string dataPathWithData = $"{Org}/{targetRepository}/instances?instanceOwnerPartyId=51001";
-            using HttpRequestMessage httpRequestMessage = new(HttpMethod.Post, dataPathWithData);
+            string dataPathWithData = $"{Org}/{StatefulApp}/instances/{PartyId}/{InstanceGuId}";
+            using HttpRequestMessage httpRequestMessage = new(HttpMethod.Get, dataPathWithData);
             httpRequestMessage.Headers.Referrer = new Uri($"{MockedReferrerUrl}?org={Org}&app={StatefulApp}&selectedLayoutSetInEditor={LayoutSetName}");
 
             using HttpResponseMessage response = await HttpClient.Value.SendAsync(httpRequestMessage);
@@ -54,7 +47,8 @@ namespace Designer.Tests.Controllers.PreviewController
             string responseBody = await response.Content.ReadAsStringAsync();
             JsonDocument responseDocument = JsonDocument.Parse(responseBody);
             Instance instance = JsonConvert.DeserializeObject<Instance>(responseDocument.RootElement.ToString());
-            Assert.Equal("test-datatask-id", instance.Data[0].Id);
+            Assert.Equal($"{PartyId}/{InstanceGuId}", instance.Id);
+            Assert.Equal("ttd", instance.Org);
             Assert.Equal("Task_1", instance.Process.CurrentTask.ElementId);
         }
     }
